use impl method
